redesign the database structure and make the needed modules
Merging into master here becasue we have a working version of the base
functionality at this point. In each module there are still some things
to be done, but the list, create and delete is working for each table.

Probably holding development on here for now, first gonna try and throw
together a frontend. This will also greatly improve the testing for this
backend.